Use the same port everywhere
Steam (or CS:GO or Valve) isn't too happy about using a different port when
running the CS:GO server and then use some kind of port forwarding. But only
when connecting from the internet. Locally it doesn't seem to be an issue..
So this should hopefully enable internet access to our servers.